billing: extract Stripe refund helpers
Move Stripe refund-specific lookup and payment state logic into a dedicated refunds module so the webhook and payment sync diffs are easier to review.

Made-with: Cursor
